+++ test/data/talker/error_classifier_test.dart
@@ -5,8 +5,8 @@ import 'package:dio/dio.dart';
 import 'package:domain_error/domain_error.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:observatory/src/data/error_classifier.dart';
-import 'package:observatory/src/domain/_barrel.dart';
+import 'package:observatory/src/feature/_common/domain/_barrel.dart';
+import 'package:observatory/src/feature/_common/infrastructure/error_classifier.dart';
 
 void main() {
   const classifier = FrameworkErrorClassifier();