+++ test/data/talker/managed_talker_test.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 
 import 'package:flutter_test/flutter_test.dart';
 import 'package:observatory/observatory.dart';
-import 'package:observatory/src/data/log_sanitizer.dart';
+import 'package:observatory/src/feature/_common/infrastructure/log_sanitizer.dart';
 import 'package:talker/talker.dart' as talker;
 
 import '../../support.dart';